requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.121.2
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.1.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.2.2
pillow==11.3.0
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
pypdf==6.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.4
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uvicorn==0.38.0
xgboost==3.0.5
